LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-option-element/option-text-label-expected.txt


FAIL Option with non-empty label. assert_equals: expected "text" but got "label"
PASS Option with empty label. 
